Change from WebSocket to REST API
